+++ web/usuaris/src/__mocks__/handlers.ts
@@ -6,11 +6,38 @@ export const handlers = [
     HttpResponse.json(
       [
         {
-          id: 'TestId',
-          name: 'Test Name',
-          status: 'INACTIVE',
-          createdAt: '2023-11-15T15:36:02.234Z',
-          itineraryName: 'backend-node',
+          id: '1',
+          name: 'Ona Sitgar',
+          status: 'PENDING',
+          createdAt: '2023/11/05 00:00:00.000',
+          itineraryName: 'Backend Node',
+        },
+        {
+          id: '2',
+          name: 'Marc Bofill',
+          status: 'ACTIVE',
+          createdAt: '2023/11/05 00:00:00.000',
+          itineraryName: 'Frontend React',
+        },
+        {
+          id: '3',
+          name: 'Montserrat Capdevila',
+          status: 'BLOCKED',
+          createdAt: '2023/11/05 00:00:00.000',
+          itineraryName: 'Fullstack Php',
+        },
+      ],
+      { status: 200 }
+    )
+  ),
+  
+  http.get(urls.getItineraries, () =>
+    HttpResponse.json(
+      [
+        {
+          id: '1',
+          name: 'Frontend React',
+          slug: 'react',
         },
       ],
       { status: 200 }
@@ -38,6 +65,9 @@ export const errorHandlers = [
     HttpResponse.json({ message: 'Invalid Credentials' }, { status: 401 })
   ),
   http.get(urls.getUsers, () =>
+    HttpResponse.json({ message: 'Database error' }, { status: 500 })
+  ),
+  http.get(urls.getItineraries, () =>
     HttpResponse.json({ message: 'Database error' }, { status: 500 })
   ),
 ]