+++ web/usuaris/src/constants/urls.ts
@@ -1,4 +1,5 @@
 export const urls = {
+  getItineraries: '/api/v1/itineraries',
   getUsers: '/api/v1/dashboard/users',
   logIn: '/api/v1/dashboard/auth/login',
   getMe: '/api/v1/dashboard/users/me',