+++ web/usuaris/src/types/index.tsx
@@ -1,1 +1,2 @@
 export { type TUserData } from './types'
+export { type TItinerary } from './types'